Refactor: Professional Architectural Hardening (Phase 15)
Completed Phase 15: Professional Architectural Hardening, ensuring Sniper acts as a modern, extensible, and high-performance research platform.

Fine-grained Synchronization & Scalability:
- Refactored SyncServer, SyscallServer, and BarrierSyncServer to use per-resource locks.
- Reduced reliance on the global ThreadManager lock to improve simulation KIPS, implementing per-thread locking for thread stalls/resumes and a lock-free atomic active thread count.
- Implemented RecursiveLock and applied it to the Scheduler hierarchy to prevent deadlocking in hook dispatches.

Extensibility & Component Registry:
- Implemented a static ComponentRegistry factory template.
- Removed hardcoded switch statements for component instantiation, migrating BranchPredictor and CacheSet to self-registering modules.
- Enforced strict decoupling by removing concrete implementation headers from base classes.

Miscellaneous:
- Handled missing Sift::Mode enum values (ModeCacheOnly, ModeNoSimulation) in the frontend control switch statement.
- Fixed initialization order warnings (-Wreorder) in ThreadManager and CoreManager.
- Checked off completed tasks in GEMINI.md.

+++ common/scheduler/scheduler_big_small.cc
@@ -1,11 +1,11 @@
 #include "scheduler_big_small.h"
-#include "simulator.h"
 #include "config.hpp"
-#include "thread.h"
-#include "performance_model.h"
 #include "core_manager.h"
 #include "misc/tags.h"
+#include "performance_model.h"
 #include "rng.h"
+#include "simulator.h"
+#include "thread.h"
 
 // Example random big-small scheduler using thread affinity
 //
@@ -20,27 +20,24 @@
 // of SchedulerPinnedBase
 
 SchedulerBigSmall::SchedulerBigSmall(ThreadManager *thread_manager)
-   : SchedulerPinnedBase(thread_manager, SubsecondTime::NS(Sim()->getCfg()->getInt("scheduler/big_small/quantum")))
-   , m_debug_output(Sim()->getCfg()->getBool("scheduler/big_small/debug"))
-   , m_last_reshuffle(SubsecondTime::Zero())
-   , m_rng(rng_seed(42))
+    : SchedulerPinnedBase(thread_manager, SubsecondTime::NS(Sim()->getCfg()->getInt("scheduler/big_small/quantum"))),
+      m_debug_output(Sim()->getCfg()->getBool("scheduler/big_small/debug")), m_last_reshuffle(SubsecondTime::Zero()),
+      m_rng(rng_seed(42))
 {
-   // Figure out big and small cores, and create affinity masks for the set of big cores and the set of small cores, respectively
+   // Figure out big and small cores, and create affinity masks for the set of big cores and the set of small cores,
+   // respectively
 
    m_num_big_cores = 0;
    CPU_ZERO(&m_mask_big);
    CPU_ZERO(&m_mask_small);
 
-   for (core_id_t coreId = 0; coreId < (core_id_t) Sim()->getConfig()->getApplicationCores(); coreId++)
-   {
+   for (core_id_t coreId = 0; coreId < (core_id_t)Sim()->getConfig()->getApplicationCores(); coreId++) {
       bool isBig = Sim()->getTagsManager()->hasTag("core", coreId, "big");
-      if (isBig)
-      {
+      if (isBig) {
          ++m_num_big_cores;
          CPU_SET(coreId, &m_mask_big);
       }
-      else
-      {
+      else {
          CPU_SET(coreId, &m_mask_small);
       }
    }
@@ -59,8 +56,7 @@ void SchedulerBigSmall::threadStall(thread_id_t thread_id, ThreadManager::stall_
    if (m_debug_output)
       std::cout << "[SchedulerBigSmall] thread " << thread_id << " stalled" << std::endl;
 
-   if (m_thread_isbig[thread_id])
-   {
+   if (m_thread_isbig[thread_id]) {
       // Pick a new thread to run on the big core(s)
       pickBigThread();
       // Move this thread to the small core(s)
@@ -81,8 +77,7 @@ void SchedulerBigSmall::threadExit(thread_id_t thread_id, SubsecondTime time)
    if (m_debug_output)
       std::cout << "[SchedulerBigSmall] thread " << thread_id << " ended" << std::endl;
 
-   if (m_thread_isbig[thread_id])
-   {
+   if (m_thread_isbig[thread_id]) {
       // Pick a new thread to run on the big core(s)
       pickBigThread();
    }
@@ -98,17 +93,15 @@ void SchedulerBigSmall::periodic(SubsecondTime time)
 {
    bool print_state = false;
 
-   if (time > m_last_reshuffle + m_quantum)
-   {
+   if (time > m_last_reshuffle + m_quantum) {
       // First move all threads back to the small cores
-      for(thread_id_t thread_id = 0; thread_id < (thread_id_t)Sim()->getThreadManager()->getNumThreads(); ++thread_id)
+      for (thread_id_t thread_id = 0; thread_id < (thread_id_t)Sim()->getThreadManager()->getNumThreads(); ++thread_id)
       {
          if (m_thread_isbig[thread_id])
             moveToSmall(thread_id);
       }
       // Now promote as many threads to the big core pool as there are big cores
-      for(UInt64 i = 0; i < std::min(m_num_big_cores, Sim()->getThreadManager()->getNumThreads()); ++i)
-      {
+      for (UInt64 i = 0; i < std::min((UInt64)m_num_big_cores, (UInt64)Sim()->getThreadManager()->getNumThreads()); ++i) {
          pickBigThread();
       }
 
@@ -120,7 +113,7 @@ void SchedulerBigSmall::periodic(SubsecondTime time)
    SchedulerPinnedBase::periodic(time);
 
    if (print_state && m_debug_output)
-         printState();
+      printState();
 }
 
 void SchedulerBigSmall::moveToSmall(thread_id_t thread_id)
@@ -141,16 +134,13 @@ void SchedulerBigSmall::pickBigThread()
 
    // First build a list of all eligible cores
    std::vector<thread_id_t> eligible;
-   for(thread_id_t thread_id = 0; thread_id < (thread_id_t)Sim()->getThreadManager()->getNumThreads(); ++thread_id)
-   {
-      if (m_thread_isbig[thread_id] == false && m_threads_runnable[thread_id])
-      {
+   for (thread_id_t thread_id = 0; thread_id < (thread_id_t)Sim()->getThreadManager()->getNumThreads(); ++thread_id) {
+      if (m_thread_isbig[thread_id] == false && m_threads_runnable[thread_id]) {
          eligible.push_back(thread_id);
       }
    }
 
-   if (eligible.size() > 0)
-   {
+   if (eligible.size() > 0) {
       // Randomly select a thread from our list
       thread_id_t thread_id = eligible[rng_next(m_rng) % eligible.size()];
       moveToBig(thread_id);
